refactor : 불변 객체를 사용하여 LoggingContext의 캡슐화 및 데이터 무결성을 보장하도록 변경
