+++ detector/src/main/java/io/jeyong/detector/context/LoggingContext.java
@@ -1,13 +1,17 @@
 package io.jeyong.detector.context;
 
-import io.jeyong.detector.dto.RequestLogDto;
+import java.util.Map;
 
 public final class LoggingContext {
 
     private final ThreadLocal<RequestLogDto> currentLoggingForm = ThreadLocal.withInitial(RequestLogDto::new);
 
-    public RequestLogDto getCurrentLoggingForm() {
-        return currentLoggingForm.get();
+    public void logQueryOccurrence(final String query) {
+        currentLoggingForm.get().addQueryOccurrence(query);
+    }
+    
+    public Map<String, Long> getCurrentQueryOccurrences() {
+        return currentLoggingForm.get().getNPlusOneQueries();
     }
 
     public void clearLoggingContext() {
