+++ detector/src/main/java/io/jeyong/detector/context/RequestLogDto.java
@@ -1,5 +1,6 @@
-package io.jeyong.detector.dto;
+package io.jeyong.detector.context;
 
+import java.util.Collections;
 import java.util.HashMap;
 import java.util.Map;
 
@@ -7,11 +8,12 @@ public final class RequestLogDto {
 
     private final Map<String, Long> queryOccurrences = new HashMap<>();
 
-    public void addQueryOccurrence(final String query) {
+    void addQueryOccurrence(final String query) {
         queryOccurrences.put(query, queryOccurrences.getOrDefault(query, 0L) + 1);
     }
 
     public Map<String, Long> getNPlusOneQueries() {
-        return queryOccurrences;
+        // Using Collections.unmodifiableMap for Java 8 compatibility
+        return Collections.unmodifiableMap(queryOccurrences);
     }
 }
